A pre alpha version of the pluginsystem in albert.
The application is now based on plugins. This commit is for versioning
history and easy debugging only. Former services are not implemented.
The only service (now called extension) is the applauncher. Settings are
messed up. Plugin configuration is not implemented yet. No potential
dangling pointers anymore, since the extension do pass IDs (QString).