Let itemhandling be done by shared_ptr
The backend is now pretty complete. QueryResults are now called items. The items
passed around are shared pointers. This combines the advantages of the items by
id and items by pointer. Items are safe to acces in any situation like by ids
and unlike by raw pointers and efficient like raw pointers and unlike the ids.
Add a stub for language extensions.