Prototype Revamp file indexing. Closes #443
Add a tree structured folder index for smart indexing an less space
requirements. QStrings are expensive and substrings of the path repeat
itself. Build a path three to link the files against, for cheap path
storage. Further the tree can store lastModified times to determine if
an update is necessary.

Add new file types the regular file is now StandardFile class the
IndexFile class is the file linked to the index tree.

This is not done yet. Serialization is missing.