docs: Add custom docset logic
This is a bit gross in order to avoid changing the existing code and
indentation, so the patch will apply for longer. But essentially, once
we've read the network/cached docset list, we then open a custom list,
read and parse that, and re-run the logic so that we can add in the
custom docsets too.

Signed-off-by: Stephen Brennan <stephen@brennan.io>