[widgetsboxmodel] Eventbased statemachine approach
Allows for customizations of keys and such. Also avoids the transistion
explosion. This is an enourmous benefit since statemachine coding is a real pain.
Also makes the machine code _much_ more readable and maintainable. Also contains
a fix for this annoying "Qt does not send enter and leave events on window de/
activation" problem.